+++ src/components/Accordion/styles.ts
@@ -1,4 +1,4 @@
-import { Theme } from '@material-ui/core/styles'
+import { Theme, createStyles } from '@material-ui/core/styles'
 
 import { createPropertiesStyles } from '../styles'
 import { PicassoProvider } from '../Picasso'
@@ -28,7 +28,7 @@ export default ({ palette, typography }: Theme) => {
     position: 'absolute'
   })
 
-  return {
+  return createStyles({
     root: {
       background: 'transparent',
 
@@ -60,7 +60,17 @@ export default ({ palette, typography }: Theme) => {
     }),
     expandIcon: {
       fontSize: '0.7em',
-      color: palette.primary.main
+      color: palette.grey.dark,
+      transition: 'transform 150ms cubic-bezier(0.4, 0, 0.2, 1) 0ms'
+    },
+    expandIconExpanded: {
+      transform: 'rotate(180deg)'
+    },
+    expandIconAlignTop: {
+      display: 'flex',
+      alignItems: 'center',
+      height: '1.5em',
+      alignSelf: 'flex-start'
     },
     details: {
       padding: 0,
@@ -71,7 +81,11 @@ export default ({ palette, typography }: Theme) => {
       marginBottom: '0.75em'
     },
     content: {
-      display: 'block'
+      display: 'flex',
+      flexDirection: 'row',
+      alignItems: 'center',
+      justifyContent: 'space-between',
+      lineHeight: '1.5em'
     }
-  }
+  })
 }