+++ src/components/Page/story/index.jsx
@@ -4,6 +4,7 @@ import pageHeaderStory from '@components/PageHeader/story'
 import pageHeaderMenuStory from '@components/PageHeaderMenu/story'
 import pageContentStory from '@components/PageContent/story'
 import pageFooterStory from '@components/PageFooter/story'
+import pageBannerStory from '@components/PageBanner/story'
 
 import { Page } from '../Page'
 
@@ -16,6 +17,7 @@ page
   .addComponentDocs(pageHeaderMenuStory.componentDocs)
   .addComponentDocs(pageContentStory.componentDocs)
   .addComponentDocs(pageFooterStory.componentDocs)
+  .addComponentDocs(pageBannerStory.componentDocs)
 
 page
   .createChapter()
@@ -26,9 +28,12 @@ page
   })
   .addExample('Page/story/FullWidth.example.jsx', 'Full width')
   .addExample('Page/story/Scroll.example.jsx', 'Scroll with overflow')
+  .addExample('Page/story/WithBanner.example.jsx', 'With Banner')
 
 page.connect(pageHeaderStory.chapter)
 
 page.connect(pageContentStory.chapter)
 
 page.connect(pageFooterStory.chapter)
+
+page.connect(pageBannerStory.chapter)