+++ src/components/PageBanner/story/Default.example.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Page, Container } from '@toptal/picasso'
+
+const PageBannerDefaultExample = () => (
+  <Container bottom='small'>
+    <Page.Banner>
+      We are now in the process of reviewing your profile. After your profile
+      has been checked, we will reach to you via email about next steps.
+    </Page.Banner>
+  </Container>
+)
+
+export default PageBannerDefaultExample